+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 * Support Jobs lifecycle enhancements (new operations `submit`, `run`, `cancel`) ([GH-196](https://github.com/ystia/yorc/issues/196))
 * Forbid the parallel execution of several scheduled actions. This is for instance used for the asynchronous run operation of Jobs. This will prevent a same action to be scheduled in parallel (for jobs it will prevent checking and doing same actions several times) ([GH-230](https://github.com/ystia/yorc/issues/230))
+* Generate Alien 2.1-compatible events ([GH-148](https://github.com/ystia/yorc/issues/148))
+
+### BUG FIXES
+
+* Bootstrap of HA setup fails on GCP, at step configuring the NFS Client component [GH-218](https://github.com/ystia/yorc/issues/218))
+* Publish workflow events when custom workflow is finished [GH-234](https://github.com/ystia/yorc/issues/234))
 
 ## 3.1.0-M7 (December 07, 2018)